requirements.txt
Flask==3.0.0
PyMuPDF==1.24.0
Pillow==10.2.0
Werkzeug==3.0.1
requests==2.31.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.123.0